refactor(web): extract App pure logic into lib modules (CS-6 PR4)
App.tsx held ~200 lines of pure logic mixed into the component: route
parsing, scan-status formatting, and live-session-update merging. Extract
them into three lib modules so they're independently testable.

- lib/view-state.ts: parseViewState + ViewState type (route → view state)
- lib/scan-format.ts: formatScanStatusLabel, formatAgentScanProgress,
  formatIsoDate, formatWindowLabel, formatSearchSubtitle,
  getAgentDisplayCount, formatRelativeTime
- lib/live-update.ts: applyLiveSessionUpdate, compareSessionActivityDesc

Each module gets a test file (21 new unit tests). App.tsx: -204 lines.

Refs CS-6 (multi-PR; PR4 of 5).
